+++ src/base/admin.py
@@ -1,0 +1,8 @@
+from django.contrib.admin import ModelAdmin, register
+
+from base.models import LegalAndPrivacy
+
+
+@register(LegalAndPrivacy)
+class LegalAndPrivacyAdmin(ModelAdmin):
+    pass