+++ src/base/translation.py
@@ -1,0 +1,11 @@
+from modeltranslation.translator import TranslationOptions, register
+
+from base.models import LegalAndPrivacy
+
+
+@register(LegalAndPrivacy)
+class LegalAndPrivacyTranslationOptions(TranslationOptions):
+    fields = (
+        "title",
+        "text",
+    )